moved data fetch to a resouces file
