+++ client/package.json
@@ -9,7 +9,6 @@
 	},
 	"dependencies": {
 		"@chakra-ui/core": "^1.0.0-rc.4",
-		"@headlessui/react": "^0.2.0",
 		"axios": "^0.20.0",
 		"framer-motion": "^2.7.5",
 		"graphql": "^15.3.0",
